refactor: refactor the ui components in logical folders
