+++ app/page.tsx
@@ -1,5 +1,5 @@
 import styles from './page.module.css';
-import GameContainer from '@/app/ui/game/game-container';
+import GameContainer from '@/app/ui/game/container';
 
 export default function Home() {
   return <GameContainer />;
