+++ app/ui/board/coin.module.css
