+++ app/ui/board/status.module.css
@@ -13,7 +13,4 @@
 .status {
   font-size: 1rem;
   font-weight: 500;
-
-  /* height: 1.75rem;
-  line-height: 1.75rem; */
 }
